Add TAJAXException trait to reuse the throwException method
See #2825
